Move infra env form buttons to the bottom
